requirements.txt
google-adk>=0.1.0
google-cloud-bigquery>=3.0.0
python-dotenv>=0.19.0
fastmcp>=2.0.0
streamlit>=1.29.0
google-generativeai>=0.8.5